+++ src/feature/CardRegister/components/InfoInputSection/NumberField.tsx
@@ -1,5 +1,5 @@
-import Input from '../../../../common/components/Input/Input';
-import Label from '../../../../common/components/Label/Label';
+import Input from '../../../../common/components/Input';
+import Label from '../../../../common/components/Label';
 import styled from 'styled-components';
 import useFieldValidation from '../../../../common/hooks/useFieldValidation';
 import { isWithinMaxLength, isNumeric } from '../../utils/validator';
