+++ src/stories/feature/CardRegister/components/InfoInputSection/FieldSection.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 
-import Input from '../../../../../common/components/Input/Input';
+import Input from '../../../../../common/components/Input';
 import FieldSection from '../../../../../feature/CardRegister/components/InfoInputSection/FieldSection';
 
 const meta = {
